Relocate `BootstrapArtefactHandler` `BootstrapArtefactHandlerSpec` to grace-boot
* Move `GrailsBootstrapClass` `DefaultGrailsBootstrapClass` from grace-web to grace-boot
* Update `GrailsApplicationBuilder`
