+++ grace-test-support/src/main/groovy/org/grails/testing/GrailsApplicationBuilder.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2024 the original author or authors.
+ * Copyright 2016-2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -97,7 +97,7 @@ class GrailsApplicationBuilder {
             catch (Throwable ignore) {
             }
             try {
-                Class gcu = Class.forName('org.grails.web.servlet.context.GrailsConfigUtils')
+                Class gcu = Class.forName('org.grails.boot.web.servlet.GrailsConfigUtils')
                 gcu.configureServletContextAttributes(servletContext, grailsApplication,
                         mainContext.getBean(GrailsPluginManager.BEAN_NAME, GrailsPluginManager), mainContext)
             }
