--- grace-web/src/main/resources/META-INF/grails.factories
@@ -1,1 +1,0 @@
-grails.core.ArtefactHandler=org.grails.web.servlet.boostrap.BootstrapArtefactHandler
